refactor: 인증 관련 클래스들은 auth 및 common 패키지로 분배
