+++ backend/src/main/java/com/woowacourse/kkogkkog/auth/presentation/LoginInterceptor.java
@@ -2,7 +2,7 @@ package com.woowacourse.kkogkkog.auth.presentation;
 
 import static org.hibernate.validator.internal.metadata.core.ConstraintHelper.PAYLOAD;
 
-import com.woowacourse.kkogkkog.auth.application.JwtTokenProvider;
+import com.woowacourse.kkogkkog.auth.support.JwtTokenProvider;
 import com.woowacourse.kkogkkog.auth.exception.UnauthenticatedTokenException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
