+++ backend/src/main/java/com/woowacourse/kkogkkog/coupon/domain/Coupon.java
@@ -1,7 +1,7 @@
 package com.woowacourse.kkogkkog.coupon.domain;
 
 import com.woowacourse.kkogkkog.coupon.exception.SameSenderReceiverException;
-import com.woowacourse.kkogkkog.domain.BaseEntity;
+import com.woowacourse.kkogkkog.common.domain.BaseEntity;
 import com.woowacourse.kkogkkog.member.domain.Member;
 import javax.persistence.Column;
 import javax.persistence.Entity;
