+++ backend/src/test/java/com/woowacourse/kkogkkog/auth/support/JwtTokenProviderTest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.kkogkkog.auth.application;
+package com.woowacourse.kkogkkog.auth.support;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
