+++ backend/src/test/java/com/woowacourse/kkogkkog/support/documenation/DocumentTest.java
@@ -5,7 +5,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.woowacourse.kkogkkog.auth.application.AuthService;
-import com.woowacourse.kkogkkog.auth.application.JwtTokenProvider;
+import com.woowacourse.kkogkkog.auth.support.JwtTokenProvider;
 import com.woowacourse.kkogkkog.member.application.MemberService;
 import com.woowacourse.kkogkkog.coupon.application.CouponService;
 import com.woowacourse.kkogkkog.reservation.application.ReservationService;
